+++ docs/src/markdown/about/changelog.md
@@ -11,7 +11,7 @@
     or LCD Jab spaces. The non-standard Jab is still available via `coloraide.spaces.cam16.CAM16`, but it is no longer
     available in `coloraide.everything` and will be removed at a future time.
 -   **FIX**: The oRGB color space should be gamut mapped in `srgb` as it is a transform of the RGB space.
--   **FIX**: Okhsl and Okhsv must be clipped in sRGB.
+-   **FIX**: Okhsl and Okhsv have a rough sRGB approximation and are instead gamut mapped to their own gamut.
 -   **FIX**: Much more accurate ICtCp matrices.
 -   **FIX**: Fix typing of deeply nested arrays in `algebra`.
 -   **FIX**: Fix issue with HCT undefined channel resolver.